% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_cancer_sample.R
\docType{data}
\name{cancer_sample}
\alias{cancer_sample}
\title{This is the dataset to be used in several examples and tests in my package}
\format{
A tibble with 569 observations (rows) and 32 variables (columns)
}
\source{
\url{https://github.com/UBC-MDS/datateachr}
}
\usage{
cancer_sample
}
\description{
The dataset contains data on properties of cancer samples such as radius, area, and texture
}
\keyword{datasets}
