% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{This is the dataset to be used in several examples and tests in my package}
\format{
A tibble with 1704 observations (rows) and 6 variables (columns)
}
\source{
\url{http://www.gapminder.org/data/}
}
\usage{
gapminder
}
\description{
The dataset contains data of life expectancy, GDP per capita, and population by country
}
\keyword{datasets}
