% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupagg.R
\name{plot_with_stat}
\alias{plot_with_stat}
\title{Creating boxplot with some stats}
\usage{
plot_with_stat(data, grouping, agg)
}
\arguments{
\item{agg}{}
}
\value{

}
\description{
Creating boxplot with some stats
}
\examples{
mtcars \%>\% plot_with_stat("cyl", "mpg")
}
