% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.function.R
\name{summary_function}
\alias{summary_function}
\title{Summary function that bundles group_by function and summarize function to compute summary statistics}
\usage{
summary_function(data, col, var, na.rm = FALSE)
}
\arguments{
\item{data}{dataset which includes all required variables for the function}

\item{col}{column name by which other variables will be grouped by}

\item{var}{variable for which we want to obtain summary statistics}

\item{na.rm}{set default as FALSE and can be specified with TRUE as needed}
}
\value{
the function will return a tibble including the count of the variable per group of the col param, as well as columns for each summary statistics in the       #' function per group.
}
\description{
Function that takes inputs of your dataset of interest, column by which you want to group data, and variable for which you want to run summary statistics #' and produces summary statistics. I have also added a custom error message if the inputted variable is not numeric.
}
\examples{
summary_function(cancer_sample, diagnosis, radius_mean) #expect summary tibble for radius_mean per diagnosis group
summary_function(gapminder, continent, lifeExp) #expect summary tibble of lifeExp per continent group

}
