% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_vancouver_trees.R
\docType{data}
\name{vancouver_trees}
\alias{vancouver_trees}
\title{This is the dataset to be used in several examples and tests in my package}
\format{
A tibble with 146611 observations (rows) and 20 variables (columns)
}
\source{
\url{https://github.com/UBC-MDS/datateachr}
}
\usage{
vancouver_trees
}
\description{
The dataset includes information about public trees in Vancouver
}
\keyword{datasets}
